+++ src/Symfony/Bridge/Doctrine/Form/EventListener/MergeDoctrineCollectionListener.php
@@ -27,15 +27,31 @@ use Symfony\Component\EventDispatcher\EventSubscriberInterface;
  */
 class MergeDoctrineCollectionListener implements EventSubscriberInterface
 {
+    // Keeps BC. To be removed in 4.0
+    private $bc = true;
+
     public static function getSubscribedEvents()
     {
         // Higher priority than core MergeCollectionListener so that this one
         // is called before
-        return array(FormEvents::SUBMIT => array('onBind', 10));
+        return array(
+            FormEvents::SUBMIT => array(
+                // BC
+                array('onBind', 10),
+                array('onSubmit', 5),
+            ),
+        );
     }
 
-    public function onBind(FormEvent $event)
+    public function onSubmit(FormEvent $event)
     {
+        // If onBind() is overridden then logic has been executed
+        if ($this->bc) {
+            @trigger_error('The onBind() method is deprecated since version 3.1 and will be removed in 4.0. Use the onSubmit() method instead.', E_USER_DEPRECATED);
+
+            return;
+        }
+
         $collection = $event->getForm()->getData();
         $data = $event->getData();
 
@@ -43,6 +59,19 @@ class MergeDoctrineCollectionListener implements EventSubscriberInterface
         // performance on persistent collections
         if ($collection instanceof Collection && count($data) === 0) {
             $collection->clear();
+        }
+    }
+
+    /**
+     * Alias of {@link onSubmit()}.
+     *
+     * @deprecated since version 3.1, to be removed in 4.0.
+     *             Use {@link onSubmit()} instead.
+     */
+    public function onBind()
+    {
+        if (__CLASS__ === get_class($this)) {
+            $this->bc = false;
         }
     }
 }